Optimize performance with dynamic imports, bundle splitting, and lazy loading
