+++ app/(main)/ai-cover-letter/_components/cover-letter-preview.jsx
@@ -1,7 +1,19 @@
 "use client";
 
 import React from "react";
-import MDEditor from "@uiw/react-md-editor";
+import dynamic from "next/dynamic";
+import { Loader2 } from "lucide-react";
+
+// Dynamic import for MDEditor to reduce initial bundle size
+const MDEditor = dynamic(() => import("@uiw/react-md-editor"), {
+  ssr: false,
+  loading: () => (
+    <div className="flex items-center justify-center h-64 bg-muted rounded-md">
+      <Loader2 className="h-6 w-6 animate-spin" />
+      <span className="ml-2">Loading preview...</span>
+    </div>
+  ),
+});
 
 const CoverLetterPreview = ({ content }) => {
   return (
