+++ app/globals.css
@@ -51,118 +51,107 @@
   --card-foreground: oklch(0.145 0 0);
   --popover: oklch(1 0 0);
   --popover-foreground: oklch(0.145 0 0);
-  --primary: oklch(0.205 0 0);
-  --primary-foreground: oklch(0.985 0 0);
-  --secondary: oklch(0.97 0 0);
-  --secondary-foreground: oklch(0.205 0 0);
-  --muted: oklch(0.97 0 0);
-  --muted-foreground: oklch(0.556 0 0);
-  --accent: oklch(0.97 0 0);
-  --accent-foreground: oklch(0.205 0 0);
-  --destructive: oklch(0.577 0.245 27.325);
-  --border: oklch(0.922 0 0);
-  --input: oklch(0.922 0 0);
-  --ring: oklch(0.708 0 0);
-  --chart-1: oklch(0.646 0.222 41.116);
-  --chart-2: oklch(0.6 0.118 184.704);
-  --chart-3: oklch(0.398 0.07 227.392);
-  --chart-4: oklch(0.828 0.189 84.429);
-  --chart-5: oklch(0.769 0.188 70.08);
-  --sidebar: oklch(0.985 0 0);
-  --sidebar-foreground: oklch(0.145 0 0);
-  --sidebar-primary: oklch(0.205 0 0);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.97 0 0);
-  --sidebar-accent-foreground: oklch(0.205 0 0);
-  --sidebar-border: oklch(0.922 0 0);
-  --sidebar-ring: oklch(0.708 0 0);
+  --primary: oklch(0.145 0 0);
+  --primary-foreground: oklch(1 0 0);
+  --secondary: oklch(0.962 0 0);
+  --secondary-foreground: oklch(0.145 0 0);
+  --muted: oklch(0.962 0 0);
+  --muted-foreground: oklch(0.464 0.004 258);
+  --accent: oklch(0.962 0 0);
+  --accent-foreground: oklch(0.145 0 0);
+  --destructive: oklch(0.65 0.2 28);
+  --destructive-foreground: oklch(1 0 0);
+  --border: oklch(0.915 0.002 286);
+  --input: oklch(0.915 0.002 286);
+  --ring: oklch(0.145 0 0);
+  --chart-1: oklch(0.60 0.07 221);
+  --chart-2: oklch(0.70 0.14 165);
+  --chart-3: oklch(0.65 0.12 81);
+  --chart-4: oklch(0.72 0.23 66);
+  --chart-5: oklch(0.45 0.05 319);
+  --sidebar: oklch(1 0 0);
+  --sidebar-foreground: oklch(0.464 0.004 258);
+  --sidebar-primary: oklch(0.145 0 0);
+  --sidebar-primary-foreground: oklch(1 0 0);
+  --sidebar-accent: oklch(0.962 0 0);
+  --sidebar-accent-foreground: oklch(0.145 0 0);
+  --sidebar-border: oklch(0.915 0.002 286);
+  --sidebar-ring: oklch(0.145 0 0);
 }
 
 .dark {
-  --background: oklch(0.145 0 0);
-  --foreground: oklch(0.985 0 0);
-  --card: oklch(0.205 0 0);
-  --card-foreground: oklch(0.985 0 0);
-  --popover: oklch(0.205 0 0);
-  --popover-foreground: oklch(0.985 0 0);
-  --primary: oklch(0.922 0 0);
-  --primary-foreground: oklch(0.205 0 0);
-  --secondary: oklch(0.269 0 0);
-  --secondary-foreground: oklch(0.985 0 0);
-  --muted: oklch(0.269 0 0);
-  --muted-foreground: oklch(0.708 0 0);
-  --accent: oklch(0.269 0 0);
-  --accent-foreground: oklch(0.985 0 0);
-  --destructive: oklch(0.704 0.191 22.216);
-  --border: oklch(1 0 0 / 10%);
-  --input: oklch(1 0 0 / 15%);
-  --ring: oklch(0.556 0 0);
-  --chart-1: oklch(0.488 0.243 264.376);
-  --chart-2: oklch(0.696 0.17 162.48);
-  --chart-3: oklch(0.769 0.188 70.08);
-  --chart-4: oklch(0.627 0.265 303.9);
-  --chart-5: oklch(0.645 0.246 16.439);
-  --sidebar: oklch(0.205 0 0);
-  --sidebar-foreground: oklch(0.985 0 0);
-  --sidebar-primary: oklch(0.488 0.243 264.376);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.269 0 0);
-  --sidebar-accent-foreground: oklch(0.985 0 0);
-  --sidebar-border: oklch(1 0 0 / 10%);
-  --sidebar-ring: oklch(0.556 0 0);
+  --background: oklch(0.086 0.006 286);
+  --foreground: oklch(0.98 0.002 286);
+  --card: oklch(0.086 0.006 286);
+  --card-foreground: oklch(0.98 0.002 286);
+  --popover: oklch(0.086 0.006 286);
+  --popover-foreground: oklch(0.98 0.002 286);
+  --primary: oklch(0.98 0.002 286);
+  --primary-foreground: oklch(0.145 0 0);
+  --secondary: oklch(0.151 0.006 286);
+  --secondary-foreground: oklch(0.98 0.002 286);
+  --muted: oklch(0.151 0.006 286);
+  --muted-foreground: oklch(0.502 0.005 286);
+  --accent: oklch(0.151 0.006 286);
+  --accent-foreground: oklch(0.98 0.002 286);
+  --destructive: oklch(0.65 0.2 28);
+  --destructive-foreground: oklch(0.98 0.002 286);
+  --border: oklch(0.151 0.006 286);
+  --input: oklch(0.151 0.006 286);
+  --ring: oklch(0.98 0.002 286);
+  --chart-1: oklch(0.70 0.14 165);
+  --chart-2: oklch(0.60 0.07 221);
+  --chart-3: oklch(0.65 0.12 81);
+  --chart-4: oklch(0.72 0.23 66);
+  --chart-5: oklch(0.45 0.05 319);
+  --sidebar: oklch(0.112 0.006 286);
+  --sidebar-foreground: oklch(0.502 0.005 286);
+  --sidebar-primary: oklch(0.98 0.002 286);
+  --sidebar-primary-foreground: oklch(0.145 0 0);
+  --sidebar-accent: oklch(0.151 0.006 286);
+  --sidebar-accent-foreground: oklch(0.98 0.002 286);
+  --sidebar-border: oklch(0.151 0.006 286);
+  --sidebar-ring: oklch(0.98 0.002 286);
 }
 
-@layer base {
-  * {
-    @apply border-border outline-ring/50;
-  }
-  body {
-    @apply bg-background text-foreground;
-  }
+* {
+  border-color: hsl(var(--border));
 }
 
+body {
+  background-color: hsl(var(--background));
+  color: hsl(var(--foreground));
+}
 
-.grid-background{
+/* Optimized grid background with better performance */
+.grid-background {
   position: fixed;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
-
-  background:linear-gradient(
-    to right,
-    rgba(255, 255, 255, 0.1) 1px,
-    transparent 1px
-  ),
-  linear-gradient(
-    to bottom,
-    rgba(255, 255, 255, 0.1) 1px,
-    transparent 1px
-  );
+  background-image: 
+    linear-gradient(rgba(255, 255, 255, 0.1) 1px, transparent 1px),
+    linear-gradient(90deg, rgba(255, 255, 255, 0.1) 1px, transparent 1px);
   background-size: 50px 50px;
   pointer-events: none;
-  z-index: -1; 
+  z-index: -1;
+  transform: translateZ(0); /* Force hardware acceleration */
+  will-change: transform;
 }
 
-.grid-background::before{
+.grid-background::before {
   content: "";
   position: absolute;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
-  background: radial-gradient(circle,transparent,rgba(0,0,0,0.9));
+  background: radial-gradient(circle, transparent, rgba(0,0,0,0.9));
+  transform: translateZ(0); /* Force hardware acceleration */
 }
 
-/* @layer utilities{
-  .gradient {
-    @apply bg-gradient-to-b from-gray-400 via-gray-200 to-gray-600  
-  }
-  .gradient-title{
-    @apply gradient font-extrabold  tracking-tighter yext-transparent bg-clip-text pb-2 pr-2
-  }
-} */
-
+/* Optimized utilities */
 @layer utilities {
   .gradient-title {
     @apply bg-gradient-to-b from-gray-400 via-gray-200 to-gray-600 font-extrabold tracking-tighter text-transparent bg-clip-text pb-2 pr-2;
@@ -173,19 +162,17 @@
   }
 }
 
-
-
-.hero-image-wrapper{
+/* Optimized hero image animations */
+.hero-image-wrapper {
   perspective: 1000px;
 }
 
-.hero-image{
+.hero-image {
   transform: rotateX(15deg) scale(1);
-  transition: 0.5s ease-out;
+  transition: transform 0.5s ease-out;
   will-change: transform;
 }
 
-.hero-image.scrolled{
+.hero-image.scrolled {
   transform: rotateX(0deg) scale(1.1) translateY(40px);
-  
 }
