+++ data/industries.js
@@ -1,287 +1,31 @@
-export const industries = [
-    {
-      id: "tech",
-      name: "Technology",
-      subIndustries: [
-        "Software Development",
-        "IT Services",
-        "Cybersecurity",
-        "Cloud Computing",
-        "Artificial Intelligence/Machine Learning",
-        "Data Science & Analytics",
-        "Internet & Web Services",
-        "Robotics",
-        "Quantum Computing",
-        "Blockchain & Cryptocurrency",
-        "IoT (Internet of Things)",
-        "Virtual/Augmented Reality",
-        "Semiconductor & Electronics",
-      ],
-    },
-    {
-      id: "finance",
-      name: "Financial Services",
-      subIndustries: [
-        "Banking",
-        "Investment Banking",
-        "Insurance",
-        "FinTech",
-        "Wealth Management",
-        "Asset Management",
-        "Real Estate Investment",
-        "Private Equity",
-        "Venture Capital",
-        "Cryptocurrency & Digital Assets",
-        "Risk Management",
-        "Payment Processing",
-        "Credit Services",
-      ],
-    },
-    {
-      id: "healthcare",
-      name: "Healthcare & Life Sciences",
-      subIndustries: [
-        "Healthcare Services",
-        "Biotechnology",
-        "Pharmaceuticals",
-        "Medical Devices",
-        "Healthcare IT",
-        "Telemedicine",
-        "Mental Health Services",
-        "Genomics",
-        "Clinical Research",
-        "Healthcare Analytics",
-        "Elder Care Services",
-        "Veterinary Services",
-        "Alternative Medicine",
-      ],
-    },
-    {
-      id: "manufacturing",
-      name: "Manufacturing & Industrial",
-      subIndustries: [
-        "Automotive",
-        "Aerospace & Defense",
-        "Electronics Manufacturing",
-        "Industrial Manufacturing",
-        "Chemical Manufacturing",
-        "Consumer Goods",
-        "Food & Beverage Processing",
-        "Textile Manufacturing",
-        "Metal Fabrication",
-        "3D Printing/Additive Manufacturing",
-        "Machinery & Equipment",
-        "Packaging",
-        "Plastics & Rubber",
-      ],
-    },
-    {
-      id: "retail",
-      name: "Retail & E-commerce",
-      subIndustries: [
-        "E-commerce Platforms",
-        "Retail Technology",
-        "Fashion & Apparel",
-        "Consumer Electronics",
-        "Grocery & Food Retail",
-        "Luxury Goods",
-        "Sports & Recreation",
-        "Home & Garden",
-        "Beauty & Personal Care",
-        "Pet Products",
-        "Specialty Retail",
-        "Direct-to-Consumer (D2C)",
-        "Department Stores",
-      ],
-    },
-    {
-      id: "media",
-      name: "Media & Entertainment",
-      subIndustries: [
-        "Digital Media",
-        "Gaming & Esports",
-        "Streaming Services",
-        "Social Media",
-        "Digital Marketing",
-        "Film & Television",
-        "Music & Audio",
-        "Publishing",
-        "Advertising",
-        "Sports Entertainment",
-        "News & Journalism",
-        "Animation",
-        "Event Management",
-      ],
-    },
-    {
-      id: "education",
-      name: "Education & Training",
-      subIndustries: [
-        "EdTech",
-        "Higher Education",
-        "Professional Training",
-        "Online Learning",
-        "K-12 Education",
-        "Corporate Training",
-        "Language Learning",
-        "Special Education",
-        "Early Childhood Education",
-        "Career Development",
-        "Educational Publishing",
-        "Educational Consulting",
-        "Vocational Training",
-      ],
-    },
-    {
-      id: "energy",
-      name: "Energy & Utilities",
-      subIndustries: [
-        "Renewable Energy",
-        "Clean Technology",
-        "Oil & Gas",
-        "Nuclear Energy",
-        "Energy Management",
-        "Utilities",
-        "Smart Grid Technology",
-        "Energy Storage",
-        "Carbon Management",
-        "Waste Management",
-        "Water & Wastewater",
-        "Mining",
-        "Environmental Services",
-      ],
-    },
-    {
-      id: "consulting",
-      name: "Professional Services",
-      subIndustries: [
-        "Management Consulting",
-        "IT Consulting",
-        "Strategy Consulting",
-        "Digital Transformation",
-        "Business Advisory",
-        "Legal Services",
-        "Accounting & Tax",
-        "Human Resources",
-        "Marketing Services",
-        "Architecture",
-        "Engineering Services",
-        "Research & Development",
-        "Business Process Outsourcing (BPO)",
-      ],
-    },
-    {
-      id: "telecom",
-      name: "Telecommunications",
-      subIndustries: [
-        "Wireless Communications",
-        "Network Infrastructure",
-        "Telecom Services",
-        "5G Technology",
-        "Internet Service Providers",
-        "Satellite Communications",
-        "Data Centers",
-        "Fiber Optics",
-        "Mobile Technology",
-        "VoIP Services",
-        "Network Security",
-        "Telecom Equipment",
-        "Cloud Communications",
-      ],
-    },
-    {
-      id: "transportation",
-      name: "Transportation & Logistics",
-      subIndustries: [
-        "Electric Vehicles",
-        "Autonomous Vehicles",
-        "Logistics & Supply Chain",
-        "Aviation",
-        "Railways",
-        "Maritime Transport",
-        "Urban Mobility",
-        "Fleet Management",
-        "Last-Mile Delivery",
-        "Warehousing",
-        "Freight & Cargo",
-        "Public Transportation",
-        "Space Transportation",
-      ],
-    },
-    {
-      id: "agriculture",
-      name: "Agriculture & Food",
-      subIndustries: [
-        "AgTech",
-        "Farming",
-        "Food Production",
-        "Sustainable Agriculture",
-        "Precision Agriculture",
-        "Aquaculture",
-        "Vertical Farming",
-        "Agricultural Biotechnology",
-        "Food Processing",
-        "Organic Farming",
-        "Plant-Based Foods",
-        "Agricultural Equipment",
-        "Indoor Farming",
-      ],
-    },
-    {
-      id: "construction",
-      name: "Construction & Real Estate",
-      subIndustries: [
-        "Commercial Construction",
-        "Residential Construction",
-        "Real Estate Development",
-        "Property Management",
-        "Construction Technology",
-        "Building Materials",
-        "Infrastructure Development",
-        "Smart Buildings",
-        "Interior Design",
-        "Facilities Management",
-        "Real Estate Technology",
-        "Sustainable Building",
-        "Urban Planning",
-      ],
-    },
-    {
-      id: "hospitality",
-      name: "Hospitality & Tourism",
-      subIndustries: [
-        "Hotels & Resorts",
-        "Restaurants & Food Service",
-        "Travel Technology",
-        "Tourism",
-        "Event Planning",
-        "Vacation Rentals",
-        "Cruise Lines",
-        "Catering",
-        "Theme Parks",
-        "Travel Agencies",
-        "Hospitality Management",
-        "Online Travel Booking",
-        "Cultural Tourism",
-      ],
-    },
-    {
-      id: "nonprofit",
-      name: "Non-Profit & Social Services",
-      subIndustries: [
-        "Charitable Organizations",
-        "Social Services",
-        "Environmental Conservation",
-        "Humanitarian Aid",
-        "Education Non-Profits",
-        "Healthcare Non-Profits",
-        "Arts & Culture",
-        "Community Development",
-        "International Development",
-        "Animal Welfare",
-        "Youth Organizations",
-        "Social Enterprise",
-        "Advocacy Organizations",
-      ],
-    },
-  ];
+// Split large industry data into smaller, lazily loaded chunks
+export const getIndustryById = async (id) => {
+  const { industries } = await import('./industries-data.js');
+  return industries.find(industry => industry.id === id);
+};
+
+export const getIndustrySubset = async (startIndex = 0, count = 10) => {
+  const { industries } = await import('./industries-data.js');
+  return industries.slice(startIndex, startIndex + count);
+};
+
+export const getAllIndustries = async () => {
+  const { industries } = await import('./industries-data.js');
+  return industries;
+};
+
+// Basic industry categories for initial load
+export const industryCategories = [
+  { id: "tech", name: "Technology", count: 13 },
+  { id: "finance", name: "Financial Services", count: 13 },
+  { id: "healthcare", name: "Healthcare & Life Sciences", count: 15 },
+  { id: "manufacturing", name: "Manufacturing", count: 12 },
+  { id: "retail", name: "Retail & E-commerce", count: 10 },
+  { id: "education", name: "Education", count: 8 },
+  { id: "energy", name: "Energy & Utilities", count: 11 },
+  { id: "consulting", name: "Consulting", count: 7 },
+  { id: "media", name: "Media & Entertainment", count: 9 },
+  { id: "transportation", name: "Transportation & Logistics", count: 8 },
+  { id: "realestate", name: "Real Estate", count: 6 },
+  { id: "nonprofit", name: "Non-Profit & Government", count: 5 },
+];
