+++ package.json
@@ -4,11 +4,14 @@
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
-    
+    "build": "npx prisma generate && next build",
+    "build:analyze": "ANALYZE=true npm run build",
     "start": "next start",
     "lint": "next lint",
     "postinstall": "prisma generate",
-    "build": "npx prisma generate && next build"
+    "audit:performance": "npm audit && npm ls --depth=0",
+    "bundle:analyze": "npx @next/bundle-analyzer",
+    "optimize": "npm run lint && npm run build:analyze"
   },
   "dependencies": {
     "@clerk/nextjs": "^6.18.5",
@@ -39,14 +42,14 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.2",
     "react-spinners": "^0.17.0",
-    "recharts": "^2.15.3",
+    "recharts": "^2.15.4",
     "sonner": "^2.0.3",
     "tailwind-merge": "^3.2.0",
     "zod": "^3.24.4"
-    
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@next/bundle-analyzer": "^15.4.3",
     "@tailwindcss/postcss": "^4.1.5",
     "autoprefixer": "^10.4.21",
     "eslint": "^9",
